# course_project/trade-bot/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Trade-bot Application",
        "title": "Trade-bot API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/orderManager/send-order": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "sendOrder to kraken futures API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orderManager"
                ],
                "summary": "SendOrder",
                "operationId": "sendOrder",
                "parameters": [
                    {
                        "description": "send order info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/krakenFuturesSDK.SendOrderArguments"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "order_id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.errResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.errResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.errResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "krakenFuturesSDK.SendOrderArguments": {
            "type": "object",
            "required": [
                "order_type",
                "side",
                "size",
                "symbol"
            ],
            "properties": {
                "cli_order_id": {
                    "type": "string"
                },
                "limit_price": {
                    "type": "number"
                },
                "order_type": {
                    "type": "string"
                },
                "reduce_only": {
                    "type": "boolean"
                },
                "side": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "stop_price": {
                    "type": "number"
                },
                "symbol": {
                    "type": "string"
                },
                "trigger_signal": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}